==> 02 Scripts/1 IC Data Import and Descriptive Analysis.ipynb <==
## Table of contents

### 01. Importing libraries
### 02. Importing data
### 03. Data Review

 # 01. Importing libraries

# Import libraries
import pandas as pd
import numpy as np
import os

# 02. Importing data

# creating shortcut for folder location
path = r'/Users/tuor/DA Projects/230321 Instacart Basket Analysis'

# creating orders data frame
df1 = pd.read_csv(os.path.join(path, '02 Data', 'Original Data', 'orders.csv'), index_col = False)

# creating products data frame
df_prods = pd.read_csv(os.path.join(path, '02 Data', 'Original Data', 'products.csv'), index_col = False)

# 03. Data Review

# looking at first 5 rows and all columns
df1.head()

# looking at first 30 rows and all columns
df1.head(30)

# looking at last 5 rows and all columns

df_prods.tail()

# orders rows and columns numbers
df1.shape

# products rows and columns numbers
df_prods.shape

# getting data types, non-null count, rows and columns and other info from info() function for products
df1.info()

# getting data types, non-null count, rows and columns and other info from info() function for products
df_prods.info()

# another way of finding columns
df1.columns

# getting count, mean, standard deviation, minimum, max, etc. for orders data frame
df1.describe()

### Reviewing a smaller set of data to evaluate columns for removal

# 1000 row data frame from orders data set
dfreview = pd.read_csv(os.path.join(path, '02 Data', 'Original Data', 'orders.csv'), nrows=1000)

# getting data types, non-null count, rows and columns and other info from info() function for products
dfreview.info()

### Importing with the columns that are desired for the import

# creating dictionary of column names desired
vars_list = ['order_id',
             'user_id',
             'order_number',
             'order_dow',
             'order_hour_of_day',
             'days_since_prior_order']

# selecting only columns listed in dictionary
df = pd.read_csv(os.path.join(path, '02 Data', 'Original Data', 'orders.csv'), usecols = vars_list)

# using info to confirm columns used
df.info()

# reviewing data
print(df.head(5))

# reviewing data
print(df_prods.head(20))

# reviewing data
print(df_prods.tail(35))

# reviewing data
print(df_prods.shape)

# getting count, mean, standard deviation, minimum, max, etc. for products data frame
df_prods.describe()

# getting data types, non-null count, rows and columns and other info from info() function for products
df_prods.info()

What’s the maximum value of the “aisle_id” column?
- 134

What’s the data type of the “department_id” column?
- int64
