==> 02 Scripts/2 Wrangling Data.ipynb <==
## Table of contents

### 01. Importing libraries
### 02. Importing data
### 03. Data Wrangling


# 01. Importing Libraries

# Import libraries
import pandas as pd
import numpy as np
import os

# 02. Importing Data

# short cut to instacart project folder
path = r'/Users/tuor/DA Projects/230321 Instacart Basket Analysis'

# data frame for orders
df_ords = pd.read_csv(os.path.join(path, '02 Data', 'Original Data', 'orders.csv'), index_col = False)

# data frame for products
df_prods = pd.read_csv(os.path.join(path, '02 Data', 'Original Data', 'products.csv'), index_col = False)

# data frame for departments
df_dep = pd.read_csv(os.path.join(path, '02 Data', 'Original Data', 'departments.csv'), index_col = False)

# 03. Data Wrangling 
### (Manipulations and Transformations)

# Dropping eval_set column from orders.csv
df_ords = df_ords.drop(columns = ['eval_set'])

# review of data information for cleaning
df_ords.describe()

#Counting missing values for orders data frame for days_since_prior_order column
df_ords['days_since_prior_order'].value_counts(dropna = False)

#Renaming column order_dow to order_day_of_week
df_ords.rename(columns = {'order_dow' : 'order_day_of_week'}, inplace = True)

#Changing order_id and user_id to string so that describe() ignores them when run
df_ords['order_id'] = df_ords['order_id'].astype('str')
df_ords['user_id'] = df_ords['user_id'].astype('str')

# confirming change of data type
df_ords.info()

# reviewing columns for data frame for departments
df_dep.head()

#Transposing the data frame for department
df_dep_t = df_dep.T

# confirming transposition
df_dep_t

#reseting the index to have the 0 for the first row where the old header is
df_dep_t.reset_index()

#creating a new header at row 0
new_header = df_dep_t.iloc[0]

#Reviewing new header
new_header

# setting up new department list
df_dep_t_new = df_dep_t[1:]

# reviewing the department list
df_dep_t_new

# Placing the header onto the colums
df_dep_t_new.columns = new_header

# confirming the header was added properly
df_dep_t_new

# Turning the department dataframe into a dictionary
data_dict = df_dep_t_new.to_dict('index')

#Reviewing dictionary
data_dict

#Using search function for finding dictionary entry
print(data_dict.get('19'))

# searching data frame for department 19
df_prods['department_id']==19

# searching data frame for department 19
df_prods[df_prods['department_id']==19]

#Creating a data frame based on if it is a snack (departmetn ID 19 from the data dictionary)
df_snacks = df_prods[df_prods['department_id']==19]

# confirming data frame is created
df_snacks.head()

##### use if you want to look at multiple departments at one time
##### df_snacks_3 = df_prods.loc[df_prods['department_id'].isin([19])]

#exporting orders data frame
#df_ords.to_csv(os.path.join(path, '02 Data', 'Prepared Data', 'orders_wrangled.csv'))

# Info to assist with first task
df_ords.info()

# Info to assist with first task
df_ords

#converting order_number data type to string ("Object" in Python)
df_ords['order_number'] = df_ords['order_number'].astype('str')

# Proof of change
df_ords.info()

#Renaming column days_since_prior_order to days_since_last_order without overwriting data frame
# df_ords.rename(columns = {'days_since_prior_order' : 'days_since_last_order'}, inplace = True)

# Proof of no change
df_ords

#Counting values for orders data frame for order_hour_of_day column
#Number at the top is the busiest hour of the day for orders
df_ords['order_hour_of_day'].value_counts(dropna = False)

# Determining value of 4 in department_id column by reviewing the data dictionary created
print(data_dict.get('4'))

# For quick reference of all departments and their ID number
# Better formatting than the data dictionary in my oppinion
df_dep_t_new

# Creating a data frame for all breakfast products
df_breakfast = df_prods.loc[df_prods['department_id'].isin([14])]

# Proof of creation
df_breakfast

# Creating a data frame for all breakfast products
df_dinner_party = df_prods.loc[df_prods['department_id'].isin([5, 20, 7, 12])]

# Proof of creation
df_dinner_party

#Assesing the size of last data frame
df_dinner_party.shape

# Confirming name of column is user_id to locate customer with user_id of 1
df_ords

# Grabbing all info on customer with user ID of 1
df_user_1 = df_ords.loc[df_ords['user_id'].isin(['1'])]

# Proof of creation
df_user_1

#Basic stats on users behavior
df_user_1.describe()

#Exporting orders data frame
df_ords.to_csv(os.path.join(path, '02 Data', 'Prepared Data', 'orders_wrangled.csv'))

#Exporting departments data frame
df_dep_t_new.to_csv(os.path.join(path, '02 Data', 'Prepared Data', 'departments_wrangled.csv'))
