==> 02 Scripts/3 Data Consistency Checks.ipynb <==
## Table of Contents

### 01. Importing Libraries

### 02. Importing Data Sets

### 03. Consistency Checks



# 01. Importing Libraries

# Import libraries
import pandas as pd
import numpy as np
import os

# 02. Importing Data Sets

# short cut to instacart project folder
path = r'/Users/tuor/DA Projects/230321 Instacart Basket Analysis'

df_prods = pd.read_csv(os.path.join(path, '02 Data', 'Original Data', 'products.csv'), index_col = False)

df_ords = pd.read_csv(os.path.join(path, '02 Data', 'Prepared Data', 'orders_wrangled.csv'), index_col = False)

df_prods.describe()

df_ords.describe()

# Create a dataframe

df_test = pd.DataFrame()

# Create a mixed type column
df_test['mix'] = ['a', 'b', 1, True]

#checking if a data frame contains any mixed type columns in data frame "df_test"

for col in df_test.columns.tolist():
  weird = (df_test[[col]].applymap(type) != df_test[[col]].iloc[0].apply(type)).any(axis = 1)
  if len (df_test[weird]) > 0:
    print (col)

df_test['mix'] = df_test['mix'].astype('str')

#Finding and summing the number of null values in Products data frame
df_prods.isnull().sum()

df_nan = df_prods[df_prods['product_name'].isnull() == True]

df_nan

#creating a new data frame with all values that are not null
df_prods_clean = df_prods[df_prods['product_name'].isnull() == False]

df_prods_clean.shape

df_prods.shape

# creating a data frame of all duplicates
df_dups = df_prods_clean[df_prods_clean.duplicated()]

df_dups

df_prods_clean_no_dups = df_prods_clean.drop_duplicates()

df_prods_clean_no_dups.shape

#Exporting products data frame
df_prods_clean_no_dups.to_csv(os.path.join(path, '02 Data', 'Prepared Data', 'products_wrangled.csv'))

# 03. Consistency Checks

df_ords.describe()

Highest number for order id match up with number of records.

Order day of the week stays within 7 days which it may be a 0-6 day count of seven explaining the 6 day max value.

Time stays within 24 hours with again the system starting at 0 resulting in the 24th hour being 23.

Given the tip of min and max values perhaps day of week and hour of day should start at 1.

Days since prior order min 0 and max of 30 feels reasonable. So does the average of 11.

#Reviewing data types in in Orders data frame
df_ords.dtypes

# summing the number of null values in Orders data frame
df_ords.isnull().sum()

There are a significant number of missing values from days_since_prior_order column.

This would make sense if a majority of all of the orders were first time orders. There would be no prior orders to measure against.

# Collecting all orders with prior order date being null
df_ords[df_ords['days_since_prior_order'].isnull() == True]

df_ords_1st_ord = df_ords[df_ords['days_since_prior_order'].isnull() == True]

df_ords_1st_ord[df_ords_1st_ord.duplicated(subset=['user_id'])]

df_ords.duplicated(subset=['user_id']).sum()

# Getting number of records in df_ords
df_ords.shape

# Adding number of nan records from days_since_prior_order to the number of duplicate user_id
3214874 + 206209

Since the number of records for duplicate user_id's adds up with the number of nan to equal the orginal orders data frame record number it is logical to say that the nan are first orders made by a user.

I would not replace these values with any markers as I do not want to change the data type to mix or interfere with the statistical summations.

No Data was changed for ords and prods was exported earlier
