==> 02 Scripts/4.1 Combining & Exporting Data (Orders and Products).ipynb <==
## Table of Contents

### 01. Importing Libraries

### 02. Importing Data Sets

### 03. Combining and Exporting Data



# 01. Importing Libraries

# Import libraries
import pandas as pd
import numpy as np
import os

# 02. Importing Data Sets

# short cut to instacart project folder
path = r'/Users/tuor/DA Projects/230321 Instacart Basket Analysis'

# Products data frame
df_prods = pd.read_csv(os.path.join(path, '02 Data', 'Original Data', 'products.csv'), index_col = False)

# Orders data frame
df_ords = pd.read_csv(os.path.join(path, '02 Data', 'Prepared Data', 'orders_wrangled.csv'), index_col = False)

# Orders/Products join table data frame
df_ords_prior = pd.read_csv(os.path.join(path, '02 Data', 'Original Data', 'order_products__prior.csv'), index_col = False)

# confirming import and propper columns
df_ords_prior.head()

# 03. Combining and Exporting Data

# examining size of data frame
df_ords_prior.shape

# examining size of data frame
df_ords.shape

# Merging df_ords and df_ords_prior
df_merged_large = df_ords.merge(df_ords_prior, on = 'order_id', indicator = True)

df_merged_large = df_merged_large.drop('Unnamed: 0', axis=1)

df_merged_large.head()

df_merged_large['_merge'].value_counts()

# merging using the outer method to assess the matching of the two data sets
df_merged_large_outer = df_ords.merge(df_ords_prior, on = 'order_id', how = 'outer', indicator = True)

# assessing the matching of the two data sets
df_merged_large_outer['_merge'].value_counts()

# exporting to pickle
df_merged_large.to_pickle(os.path.join(path, '02 Data', 'Prepared Data', 'order_products_combined.pkl'))

df_merged_large.shape
