==> 02 Scripts/4.2 Combining & Exporting Data (orders with order-product join information combined with product data).ipynb <==
## Table of Contents

### 01. Importing Libraries

### 02. Importing Data Sets

### 03. Data Wrangling

### 04. Merging products and orders_products_combined

# 01. Importing Libraries

# Import libraries
import pandas as pd
import numpy as np
import os

# 02. Importing Data Sets

# creating a path variable as a short cut for future importing of data
path = r'/Users/tuor/DA Projects/230321 Instacart Basket Analysis'

# Products data frame
df_prods = pd.read_csv(os.path.join(path, '02 Data', 'Original Data', 'products.csv'), index_col = False)

# Orders data frame
df_ords = pd.read_csv(os.path.join(path, '02 Data', 'Prepared Data', 'orders_wrangled.csv'), index_col = False)

# orders_products_combined data frame
df_orders_products_combined = pd.read_pickle(os.path.join(path, '02 Data', 'Prepared Data', 'orders_products_combined.pkl'))

# 03. Data Wrangling

# Reviewing if NaN is in the data frame
df_prods.isna().sum()

# Creating a data frame of records where product_name are NaN 
nan_df = df_prods[df_prods['product_name'].isna()]

# Confirming data frame creation
nan_df

# 04. Merging products and orders_products_combined (orders with product join info)

# confirming shape of data set
df_orders_products_combined.shape

# reviewing columns for merging with products
df_orders_products_combined.head()

# reviewing columns for merging with products
df_prods.shape

# reviewing columns for merging with products
df_prods.head()

# Merging df_prods and df_orders_products_combined
df_merged_ords_prods = df_prods.merge(df_orders_products_combined, on = 'product_id', indicator = 'merge_indicator')

#Reviewing merger
df_merged_ords_prods['merge_indicator'].value_counts()

# merging using the outer method to assess the matching of the two data sets
df_merged_ords_prods_outer = df_prods.merge(df_orders_products_combined, on = 'product_id', how = 'outer', indicator = 'merge_indicator')

#Reviewing merger
df_merged_ords_prods_outer['merge_indicator'].value_counts()

#Looking through new data frame for consistency
df_merged_ords_prods_outer.head()

# removing '_merge' column
df_merged_ords_prods_outer = df_merged_ords_prods_outer.drop('_merge', axis = 1)

# exporting to pickle
df_merged_ords_prods.to_pickle(os.path.join(path, '02 Data', 'Prepared Data', 'orders_products_merged.pkl'))
