==> 02 Scripts/5 Deriving New Variables.ipynb <==
### Table of Contents

01. Importing Libraries

02. Importing Data Sets

03. Defining Functions

04. Testing price_label Function on Practice Data Frame

05. Test loc Method to Create a Flagging Column

06. loc Method to Create a Flagging Column for Price

07. Trouble Shooting application of loc Method for Whole Data Frame

08. Updating busiest_day Column Based on Client Feedback

09. Creating busiest_period_of_day column

# 01. Importing Libraries

# Import libraries
import pandas as pd
import numpy as np
import os

# 02. Importing Data Sets

# Creating a path variable as a short cut for future importing of data
path = r'/Users/tuor/DA Projects/230321 Instacart Basket Analysis'

# Importing orders_products_merged data frame
df_ord_prod_merge = pd.read_pickle(os.path.join(path, '02 Data', 'Prepared Data', 'orders_products_merged.pkl'))

# Subset environment of orders_products_merged to test execution of edits
df = df_ord_prod_merge[:1000000]

# 03. Defining Functions

# Price Label function
def price_label(row):

  if row['prices'] <= 5:
    return 'Low-range product'
  elif (row['prices'] > 5) and (row['prices'] <= 15):
    return 'Mid-range product'
  elif row['prices'] > 15:
    return 'High range'
  else: return 'Not enough data'

# 04. Testing price_label Function on Practice Data Frame

# Applying price_label function to create a column to bucket prices
df['price_range'] = df.apply(price_label, axis=1)

# counting the values in the price_range column
df['price_range'].value_counts(dropna = False)

# Checking the max value in prices
df['prices'].max()

# 05. Test loc Method to Create a Flagging Column

# If prices > 15 using loc method to label flagging column
df.loc[df['prices'] > 15, 'price_range_loc'] = 'High-range product'

# If prices > 5 and <= 15 using loc method to label flagging column
df.loc[(df['prices'] <= 15) & (df['prices'] > 5), 'price_range_loc'] = 'Mid-range product' 

# If prices <= 5 using the loc method to label flagging column
df.loc[df['prices'] <= 5, 'price_range_loc'] = 'Low-range product'

# Counting new flagging column to confirm results
df['price_range_loc'].value_counts(dropna = False)

# 06. loc Method to Create a Flagging Column for Price

# If prices > 15 using loc method to label flagging column
df_ord_prod_merge.loc[df_ord_prod_merge['prices'] > 15, 'price_range_loc'] = 'High-range product'

# If prices > 5 and <= 15 using loc method to label flagging column
df_ord_prod_merge.loc[(df_ord_prod_merge['prices'] <= 15) & (df_ord_prod_merge['prices'] > 5), 'price_range_loc'] = 'Mid-range product' 

# If prices <= 5 using the loc method to label flagging column
df_ord_prod_merge.loc[df_ord_prod_merge['prices'] <= 5, 'price_range_loc'] = 

# Counting new flagging column to confirm results
df_ord_prod_merge['price_range_loc'].value_counts(dropna = False)

# 07. Trouble Shooting application of loc Method for Whole Data Frame

# create a boolean mask for NaN values in price_range_loc column
mask = df_ord_prod_merge['price_range_loc'].isna()

# filter the data frame using boolean indexing
df_filtered = df_ord_prod_merge.loc[mask]

df_filtered.head()

Found an error in my loc code for flagging Mid-range product where it was calling from the wrong data frame resulting in a significan't number of NaN results.

# 08. Creating busiest_day Column

df_ord_prod_merge['order_day_of_week'].value_counts(dropna = False)

# For Loop to create a flag column to note level of business
result = []

for value in df_ord_prod_merge["order_day_of_week"]:
  if value == 0:
    result.append("Busiest day")
  elif value == 4:
    result.append("Least busy")
  else:
    result.append("Regularly busy")

result

# Creating busiest_day column and filling it with result
df_ord_prod_merge['busiest_day'] = result

df_ord_prod_merge['busiest_day'].value_counts(dropna = False)

# 08. Updating busiest_day Column Based on Client Feedback

# For Loop to update a flag column to note level of business
result = []

for value in df_ord_prod_merge["order_day_of_week"]:
  if value == 0:
    result.append("Busiest day")
  elif value == 3:
    result.append("2nd Least busy") 
  elif value == 4:
    result.append("Least busy")   
  else:
    result.append("Regularly busy")

result

# Updating busiest_day column and filling it with result
df_ord_prod_merge['busiest_day'] = result

# Counting busiest_day column to check work
df_ord_prod_merge['busiest_day'].value_counts(dropna = False)

# Renaming column busiest_day to busiest_days inplace
df_ord_prod_merge.rename(columns={'busiest_day': 'busiest_days'}, inplace=True)

df_ord_prod_merge.head()

# 09. Creating busiest_period_of_day column

# Assessing the busiest hours
df_ord_prod_merge['order_hour_of_day'].value_counts(dropna = False)

# For Loop to create a flag column to note level of business by hour
result = []

for value in df_ord_prod_merge["order_hour_of_day"]:
  if value == 3:
    result.append("Fewst orders")
  elif value == 10:
    result.append("Most orders")   
  else:
    result.append("Average orders")

result

# Creating busiest_period_of_day column and filling it with result
df_ord_prod_merge['busiest_period_of_day'] = result

# Counting busiest_day column to check work
df_ord_prod_merge['busiest_period_of_day'].value_counts(dropna = False)

# exporting to pickle
df_ord_prod_merge.to_pickle(os.path.join(path, '02 Data', 'Prepared Data', '230328_order_products_merged.pkl'))
