==> 02 Scripts/9 Coding Etiquette & Excel Reporting.ipynb <==
### Table of contents

01. Importing Libraries
02. Importing Data Set
03. Dictionaries and Functions
04. Personally Identifiable information review and removal
05. Regions
    - 5.1 Assigning to a region
    - 5.2 Reviewing region by spender type
06. Customer Profiling
07. Exporting Data and Plots

# 01. Importing Libraries

import pandas as pd
import numpy as np
import os
import matplotlib.pyplot as plt
import seaborn as sns
import scipy

# 02. Importing Data Set(s)

# Creating a path variable as a short cut for future importing of data
path = r'/Users/tuor/DA Projects/230321 Instacart Basket Analysis'

ords_prods_cust = pd.read_pickle(os.path.join(path, '02 Data', 'Prepared Data', '230331_orders_products_customers_merged.pkl'))

# create a dictionary to map department names to IDs
department = {1: 'frozen', 2: 'other', 3: 'bakery', 4: 'produce', 5: 'alcohol', 6: 'international', 
                    7: 'beverages', 8: 'pets', 9: 'dry goods pasta', 10: 'bulk', 11: 'personal care',
                    12: 'meat seafood', 13: 'pantry', 14: 'breakfast', 15: 'canned goods', 16: 'dairy eggs', 
                    17: 'household', 18: 'babies', 19: 'snacks', 20: 'deli', 21: 'missing'}

crosstab = pd.crosstab(ords_prods_cust['days_since_prior_order'], ords_prods_cust['order_number'], dropna = False)

pd.options.display.max_columns = None

ords_prods_cust.head()

crosstab.head(15)

# 03. Dictionaries and Functions

# create a dictionary of regions and their states
regions = {
    'Northeast': ['Maine', 'New Hampshire', 'Vermont', 'Massachusetts', 'Rhode Island', 'Connecticut', 'New York', 'Pennsylvania', 'New Jersey'],
    'Midwest': ['Wisconsin', 'Michigan', 'Illinois', 'Indiana', 'Ohio', 'North Dakota', 'South Dakota', 'Nebraska', 'Kansas', 'Minnesota', 'Iowa', 'Missouri'],
    'South': ['Delaware', 'Maryland', 'District of Columbia', 'Virginia', 'West Virginia', 'North Carolina', 'South Carolina', 'Georgia', 'Florida', 'Kentucky', 'Tennessee', 'Mississippi', 'Alabama', 'Oklahoma', 'Texas', 'Arkansas', 'Louisiana'],
    'West': ['Idaho', 'Montana', 'Wyoming', 'Nevada', 'Utah', 'Colorado', 'Arizona', 'New Mexico', 'Alaska', 'Washington', 'Oregon', 'California', 'Hawaii']
}

# create a function to assign the region based on the state
def assign_region(state):
    for region, states in regions.items():
        if state in states:
            return region
    return 'Unknown'

# 04. Personally Identifiable information review and removal

# confirming shape of data frame
ords_prods_cust.shape

# allowing for all columns to be seen
pd.set_option('display.max_columns', None)

# reviewing columns to see if any PII exists
ords_prods_cust.head(5)

# For PII removing First and last name of the user

ords_prods_cust = ords_prods_cust.drop(columns = ['first_name', 'last_name'])

# Confirming Removale of first and last name columns
ords_prods_cust.head(5)

# 05. Regions

## 05.1 Assigning to a region

ords_prods_cust['region'] = ords_prods_cust['state'].apply(assign_region)

pd.set_option('display.max_rows', None)

# spot check to confirm regions properly defined
ords_prods_cust[['state','region']].tail(100)

# spot check to confirm regions properly defined
ords_prods_cust[['state','region']].head(100)

# reviewing assignment of region and that there are no erronious values
ords_prods_cust['region'].value_counts(dropna=False)

# reviewing assignment of region and that there are no erronious values
ords_prods_cust['state'].value_counts(dropna=False)

## 05.2 Reviewing region by spender type

# spender habit by region
ords_prods_cust.groupby('region')['spender_flag'].value_counts()

# 06. Customer Profiling

# creating list of users with greater than 5 orders
high_revenue_customers = ords_prods_cust.loc[(ords_prods_cust['max_order'] > 5)]

# confirming list removes all with max_order under 5
high_revenue_customers.value_counts('max_order').sort_index()

# exporting high revenue customers
high_revenue_customers.to_pickle(os.path.join(path, '02 Data', 'Prepared Data', '230405_high_revenue_customers.pkl'))

# New Parent: profile requirements
# Age <= 40, 
# Income > 52,200
# department 18
# dependents > 0

new_parent = high_revenue_customers.loc[(high_revenue_customers['income'] > 52200) & (high_revenue_customers['age'] <= 45) & (high_revenue_customers['number_of_dependents'] > 0) & (high_revenue_customers['department_id'] == 18)]['user_id'].unique()

# New Parent: Filling in customer profile
high_revenue_customers.loc[high_revenue_customers['user_id'].isin(new_parent), 'customer_profile'] = 'New parent'

# Young Professional: profile requirements
#Age <= 40, 
#income > 52,200, 
#dependents = 0 
#Department = irrelevant

high_revenue_customers.loc[(high_revenue_customers['income'] > 52200) & (high_revenue_customers['age'] <= 40) & (high_revenue_customers['number_of_dependents'] == 0), 'customer_profile'] = 'Young Professional'

# Pet Owner: profile requirements - will write over other profiles
#Age irrelevant
#Income > 52,200
#Department = has 8 in purchases
#Number of dependents irrelevant

pet_parents = high_revenue_customers[(high_revenue_customers['income'] > 52200) & (high_revenue_customers['department_id'] == 8)]['user_id'].unique()

# Pet Owner: Filling in customer profile - will write over other profiles
high_revenue_customers.loc[high_revenue_customers['user_id'].isin(pet_parents), 'customer_profile'] = 'Pet owner'

# Elderly: profile requirements - Will write over other profiles
#Age > 67
#Income irrelevant
#Department irrelevant
#Number of dependents irrelevant

high_revenue_customers.loc[high_revenue_customers['age'] >= 67, 'customer_profile'] = 'Elderly'

# High income adult: profile requirements - Will over write other profiles
#Age >= 35
#Income > 155,000
#Department irrelevant
#Dependents irrelevant

high_revenue_customers.loc[(high_revenue_customers['age'] >= 25) & (high_revenue_customers['income'] > 155000), 'customer_profile'] = 'High income adult'

# Other - removing remaining NaN values
high_revenue_customers['customer_profile'].fillna('Other', inplace=True)

high_revenue_customers['customer_profile'].value_counts(dropna=False)

customer_profile_counts = high_revenue_customers['customer_profile'].value_counts(dropna=True)

# Create the bar plot of order count by customer profile
cust_prof_cnt = customer_profile_counts.plot.bar()

# Set the title and axis labels
cust_prof_cnt.set_title('Count of Orders by Customer Profiles')
cust_prof_cnt.set_xlabel('Customer Profile')
cust_prof_cnt.set_ylabel('Count')
# set the figure size and DPI
cust_prof_cnt.get_figure().set_size_inches(14, 15)
cust_prof_cnt.get_figure().dpi = 100

# Group the data by 'customer_profile' and calculate the max, mean, and min of 'median_prior_order' and 'mean_user_product_price' for each group
grouped_data = high_revenue_customers.groupby('customer_profile')[['median_prior_order', 'mean_user_product_price']].agg(['max', 'mean', 'min'])

#printing values
grouped_data.head(10)

high_revenue_customers.shape

# map department names to IDs in the high_revenue_customers data frame
high_revenue_customers['department'] = high_revenue_customers['department_id'].map(department)


high_revenue_customers.shape

high_revenue_customers['department'].value_counts(dropna=False)

department_order_count = high_revenue_customers['department'].value_counts(dropna=False)

# exporting high revenue customers
high_revenue_customers.to_pickle(os.path.join(path, '02 Data', 'Prepared Data', '2304013_high_revenue_customers.pkl'))

# customer count by department unique user_id count
#customers_by_department_counts = high_revenue_customers.groupby('department')['user_id'].nunique()


# Group the data by 'customer_profile' and calculate the max, mean, and min of 'median_prior_order' and 'mean_user_product_price' for each group
grouped_data_department = high_revenue_customers.groupby('department')[['median_prior_order', 'mean_user_product_price']].agg(['max', 'mean', 'min'])

grouped_data_department.head(30)

# Create the bar plot
dep_ord_cnt = department_order_count.plot(kind='bar', figsize=(10, 14))

# Set the title and axis labels
plt.title('Count of Orders by Department')
plt.xlabel('Department')
plt.ylabel('Count')

# Set the title and axis labels
cust_prof_cnt.set_title('Count of Orders by Customer Profiles')
cust_prof_cnt.set_xlabel('Customer Profile')
cust_prof_cnt.set_ylabel('Count')
# set the figure size and DPI
cust_prof_cnt.get_figure().set_size_inches(14, 15)
cust_prof_cnt.get_figure().dpi = 100

high_revenue_customers['region'].value_counts(dropna=False)

region_order_count = high_revenue_customers['region'].value_counts(dropna=False)

# Group the data by 'customer_profile' and calculate the max, mean, and min of 'median_prior_order' and 'mean_user_product_price' for each group
grouped_data_region = high_revenue_customers.groupby('region')[['median_prior_order', 'mean_user_product_price']].agg(['max', 'mean', 'min'])

grouped_data_region.head(30)

# Create the bar plot
reg_ord_cnt = region_order_count.plot(kind='bar', figsize=(10, 10))

# Set the title and axis labels
plt.title('Count of Orders by Region')
plt.xlabel('Region')
plt.ylabel('Count')

# 07. Exporting Data and Plots

# saving customer profile by order count bar plot
cust_prof_cnt.figure.savefig(os.path.join(path, '04 Analysis','Visualizations', 'customer_profile_order_count.png'))

# saving department by order count bar plot

dep_ord_cnt.figure.savefig(os.path.join(path, '04 Analysis','Visualizations', 'department_order_count.png'))

# saving region by order count bar plot

reg_ord_cnt.figure.savefig(os.path.join(path, '04 Analysis','Visualizations', 'region_order_count.png'))
